# specs/003-develop-ui/contracts/bff-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: DevOps Career Agent Browser BFF API
  version: 1.0.0
  description: >-
    Canonical UI-to-BFF contract. Browser-facing payloads are presentation-safe
    camelCase DTOs. The BFF validates this boundary, maps to the private core
    contract, and never exposes OAuth tokens, authority claims, answer keys, or
    another employee's identifiers.
servers:
  - url: /bff/v1
security:
  - sessionCookie: []
paths:
...
  /progress/check-ins:
    post:
      operationId: createBrowserProgressCheckIn
      parameters:
        - $ref: "#/components/parameters/UiContractVersion"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/Origin"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressCheckInRequest"
      responses:
        "200":
          description: Persisted progress review with exactly one next action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressReview"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /progress/reviews:
    get:
      operationId: getBrowserProgressReview
      parameters:
        - $ref: "#/components/parameters/RoadmapIdQuery"
      responses:
        "200":
          description: Latest employee-owned persisted progress review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressReview"
        "401":
          $ref: "#/components/responses/SessionNotActive"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-learning_session
  parameters:
...
    RoadmapIdQuery:
      name: roadmapId
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    UiContractVersion:
      name: X-UI-Contract-Version
      in: header
      required: true
      schema:
        type: string
        pattern: "^1\\.[0-9]+\\.[0-9]+$"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 16
...
    Origin:
      name: Origin
      in: header
      required: true
      schema:
        type: string
        format: uri
...
  responses:
    SessionNotActive:
      description: Missing, expired, revoked, or otherwise invalid BFF session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    Forbidden:
      description: Session, Origin, CSRF, lifecycle, or owner policy denied access.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Owner-scoped resource does not exist or is not visible.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Contract-version, idempotency, or state conflict.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationFailed:
      description: Request schema, field, path, or query validation failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RateLimited:
      description: Browser actor or required dependency is temporarily throttled.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ServiceUnavailable:
      description: BFF, core, identity, session, or persistence dependency unavailable.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    GatewayTimeout:
      description: A bounded downstream operation timed out with safe retry semantics.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    ProgressCheckInRequest:
      type: object
      additionalProperties: false
      required: [roadmapId]
      properties:
        employeeProfileId:
          type: [string, "null"]
          deprecated: true
          description: Compatibility relationship only; never grants authority.
        roadmapId:
          type: string
          minLength: 1
        notes:
          type: string
          default: ""
        completedSteps:
          type: array
          description: >-
            Legacy milestone title and skill-area references retain their
            documented trim/case-insensitive all-matches behavior.
          items:
            type: string
            minLength: 1
        completedMilestoneKeys:
          type: array
          description: >-
            Preferred unambiguous stable keys. An unknown key is a 422 error.
          items:
            type: string
            pattern: "^[a-z0-9][a-z0-9._-]{2,127}$"
        newGoals:
          type: array
          items:
            type: string
            minLength: 1
...
    CareerRoadmap:
      type: object
      additionalProperties: false
      required:
        - id
        - employeeProfileId
        - goalSummary
        - currentFocus
        - milestones
        - status
        - assumptions
        - clarifyingQuestionsAsked
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        employeeProfileId:
          type: string
        previousRoadmapId:
          type: [string, "null"]
        goalSummary:
          type: string
          minLength: 1
        currentFocus:
          type: string
          minLength: 1
        milestones:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/RoadmapMilestone"
        status:
          type: string
          enum: [draft, active, revised, completed]
        assumptions:
          type: array
          items:
            type: string
        clarifyingQuestionsAsked:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoadmapMilestone:
      type: object
      additionalProperties: false
      required:
        - id
        - milestoneKey
        - ordinal
        - title
        - skillArea
        - experienceLevelFit
        - timeHorizon
        - concreteNextAction
        - reasonItMatters
        - completionState
        - priority
        - supportingNotes
      properties:
        id:
          type: string
        milestoneKey:
          type: string
          minLength: 1
        ordinal:
          type: integer
          minimum: 0
        title:
          type: string
          minLength: 1
        skillArea:
          type: string
          minLength: 1
        experienceLevelFit:
          type: string
        timeHorizon:
          type: string
          enum: [immediate, near_term, long_term]
        concreteNextAction:
          type: string
          minLength: 1
        reasonItMatters:
          type: string
        completionState:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: integer
        supportingNotes:
          type: array
          items:
            type: string
...
    ProgressReview:
      type: object
      additionalProperties: false
      required:
        - revision
        - presentation
        - followUpPrompt
        - checkIn
        - roadmapId
        - currentStatus
        - gaps
        - milestones
        - nextAction
      properties:
        revision:
          $ref: "#/components/schemas/RoadmapRevision"
        presentation:
          type: string
        followUpPrompt:
          type: string
        checkIn:
          $ref: "#/components/schemas/ProgressCheckIn"
        roadmapId:
          type: string
        currentStatus:
          type: string
        gaps:
          type: array
          items:
            type: string
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapMilestone"
        nextAction:
          $ref: "#/components/schemas/NextAction"
    RoadmapRevision:
      type: object
      additionalProperties: false
      required:
        - priorRoadmap
        - updatedRoadmap
        - completedSteps
        - normalizedMilestoneKeys
        - newGoals
        - summary
        - createdAt
      properties:
        priorRoadmap:
          $ref: "#/components/schemas/CareerRoadmap"
        updatedRoadmap:
          $ref: "#/components/schemas/CareerRoadmap"
        completedSteps:
          type: array
          items:
            type: string
        normalizedMilestoneKeys:
          type: array
          items:
            type: string
        newGoals:
          type: array
          items:
            type: string
        summary:
          type: string
        createdAt:
          type: string
          format: date-time
    ProgressCheckIn:
      type: object
      additionalProperties: false
      required:
        - id
        - employeeProfileId
        - roadmapId
        - notes
        - completedSteps
        - normalizedMilestoneKeys
        - newGoals
        - updatedRecommendations
        - createdAt
      properties:
        id:
          type: string
        employeeProfileId:
          type: string
        roadmapId:
          type: string
        notes:
          type: string
        completedSteps:
          type: array
          items:
            type: string
        normalizedMilestoneKeys:
          type: array
          items:
            type: string
        newGoals:
          type: array
          items:
            type: string
        updatedRecommendations:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapMilestone"
        createdAt:
          type: string
          format: date-time
    NextAction:
      type: object
      additionalProperties: false
      required: [kind, title, reason, target]
      properties:
        kind:
          type: string
          enum: [retry_material, resume_session, continue_milestone, review_roadmap]
        title:
          type: string
          minLength: 1
        reason:
          type: string
          minLength: 1
        target:
          type: string
          minLength: 1
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, traceId, retryable, fieldErrors]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: [string, "null"]
        code:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [field, messages]
            properties:
              field:
                type: string
              messages:
                type: array
                minItems: 1
                items:
                  type: string